+++ samples/WeihanLi.EntityFramework.Core3_0Sample/Program.cs
@@ -5,13 +5,16 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using WeihanLi.Common;
 using WeihanLi.Common.Data;
+using WeihanLi.EntityFramework.Samples;
 using WeihanLi.Extensions;
 
-namespace WeihanLi.EntityFramework.Samples
+namespace WeihanLi.EntityFramework.Core3_0Sample
 {
     public class Program
     {
-        private const string DbConnectionString = "server=.;database=Test;uid=sa;pwd=Admin888";
+        private const string DbConnectionString = @"Data Source=(localdb)\MSSQLLocalDB;Initial Catalog=master;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False"
+        //"server=.;database=TestDb;uid=sa;pwd=Admin888"
+        ;
 
         public static void Main(string[] args)
         {
@@ -23,19 +26,17 @@ namespace WeihanLi.EntityFramework.Samples
             {
                 options
                     .UseLoggerFactory(loggerFactory)
-                    .EnableDetailedErrors()
-                    .EnableSensitiveDataLogging()
-                    .UseSqlServer(DbConnectionString);
+                    //.EnableDetailedErrors()
+                    //.EnableSensitiveDataLogging()
+                    .UseSqlServer(DbConnectionString)
+                    //.AddInterceptors(new QueryWithNoLockDbCommandInterceptor())
+                    ;
             });
 
             services.AddEFRepository()
-                // .AddRepository<TestDbContext>()
                 ;
 
             DependencyResolver.SetDependencyResolver(services);
-
-            //DependencyResolver.Current.ResolveService<IEFRepositoryGenerator>()
-            //    .GenerateRepositoryCodeFor<TestDbContext>("WeihanLi.EntityFramework.Samples.Business");
 
             DependencyResolver.Current.TryInvokeService<TestDbContext>(db =>
             {
@@ -59,13 +60,23 @@ GETUTCDATE()
 
                 var abc = db.TestEntities.AsNoTracking().ToArray();
                 Console.WriteLine($"{string.Join(Environment.NewLine, abc.Select(_ => _.ToJson()))}");
-                var names = db.TestEntities.AsNoTracking().Select(t => DbFunctions.JsonValue(t.Extra, "$.Name"))
-                    .ToArray();
-                Console.WriteLine($"Names: {names.StringJoin(",")}");
+            });
+
+            DependencyResolver.Current.TryInvokeService<IEFRepositoryFactory<TestDbContext>>(repoFactory =>
+            {
+                var repo = repoFactory.GetRepository<TestEntity>();
+                var count = repo.Count();
+                Console.WriteLine(count);
             });
 
             DependencyResolver.Current.TryInvokeService<IEFRepository<TestDbContext, TestEntity>>(repo =>
             {
+                var ids0 = repo.GetResult(_ => _.Id).ToArray();
+                Console.WriteLine($"Ids: {ids0.StringJoin(",")}");
+
+                var list0 = repo.GetResult(_ => _.Id, queryBuilder => queryBuilder.WithPredict(t => t.Id > 0)).ToArray();
+                Console.WriteLine($"Ids: {list0.StringJoin(",")}");
+
                 repo.Insert(new TestEntity() { Extra = "{}", CreatedAt = DateTime.UtcNow, });
                 repo.Insert(new TestEntity() { Extra = "{}", CreatedAt = DateTime.UtcNow, });
 
@@ -101,15 +112,13 @@ GETUTCDATE()
                     .WithOrderBy(q => q.OrderByDescending(_ => _.Id)));
 
                 var list1 = repo.GetPagedListResult(x => x.Id, queryBuilder => queryBuilder
-                    .WithOrderBy(query => query.OrderByDescending(q => q.Id)), 2, 2
+                        .WithOrderBy(query => query.OrderByDescending(q => q.Id)), 2, 2
                 );
 
                 var pagedList = repo.GetPagedListResult(x => x.Id, queryBuilder => queryBuilder
-                    .WithOrderBy(query => query.OrderByDescending(q => q.Id))
-                , 1, 2);
+                        .WithOrderBy(query => query.OrderByDescending(q => q.Id))
+                    , 1, 2);
                 Console.WriteLine(pagedList.ToJson());
-
-                repo.Delete(t => DbFunctions.JsonValue(t.Extra, "$.Name") == "Abcdes");
 
                 Console.WriteLine($"Count: {repo.Count()}");
             });
@@ -121,8 +130,6 @@ GETUTCDATE()
 TRUNCATE TABLE TestEntities
 ");
             });
-
-            //
             Console.ReadLine();
         }
     }