+++ src/WeihanLi.EntityFramework/EFRepository.cs
@@ -342,9 +342,18 @@ namespace WeihanLi.EntityFramework
             return DbContext.SaveChanges();
         }
 
-        public virtual Task<TEntity> FindAsync(object[] keyValues, CancellationToken cancellationToken)
+        public virtual async Task<TEntity> FindAsync(object[] keyValues, CancellationToken cancellationToken)
         {
-            return EntitySet.FindAsync(keyValues, cancellationToken);
+            return await EntitySet.FindAsync(keyValues, cancellationToken);
+        }
+
+
+        public virtual IQueryable<TEntity> Query(Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null)
+        {
+            var queryBuilder = new EFRepositoryQueryBuilder<TEntity>(EntitySet);
+            queryBuilderAction?.Invoke(queryBuilder);
+
+            return queryBuilder.Build();
         }
 
         public virtual List<TEntity> Get(Action<EFRepositoryQueryBuilder<TEntity>> queryBuilderAction = null)