+++ src/WeihanLi.EntityFramework/IEFRepositoryFactory.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace WeihanLi.EntityFramework
+{
+    public interface IEFRepositoryFactory<TDbContext> where TDbContext : DbContext
+    {
+        IEFRepository<TDbContext, TEntity> GetRepository<TEntity>() where TEntity : class;
+    }
+}