+++ docs/ReleaseNotes.md
@@ -2,11 +2,15 @@
 
 ## WeihanLi.EntityFramework
 
+### [WeihanLi.EntityFramework 1.0.9](https://www.nuget.org/packages/WeihanLi.EntityFramework/1.0.9)
+
+- add `Delete`/`Any` for `IEFRepository`
+- update pagedList extension
+- add `DbContextBase`
 
 ### [WeihanLi.EntityFramework 1.0.8](https://www.nuget.org/packages/WeihanLi.EntityFramework/1.0.8)
 
 - rename Get with selector method name => `GetResult`/`FirstOrDefaultResult`/`GetPagedListResult`
-
 
 ### [WeihanLi.EntityFramework 1.0.7](https://www.nuget.org/packages/WeihanLi.EntityFramework/1.0.7)